pyproject.toml

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "contractor"
version = "0.1.0"
description = "Contract automation toolkit: LaTeX templates, versioning, amounts, and reports"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{name = "You"}]

dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.1",
  "Jinja2>=3.1.4",
  "sqlmodel>=0.0.22",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.3.4",
  "PyYAML>=6.0.2",
  "pypandoc>=1.13",
  "GitPython>=3.1.43",
  "Babel>=2.15.0",
  "python-moneyed>=3.0",
  "num2words>=0.5.13",
  "cn2an>=0.5.22",
  "pandas>=2.2.2",
  "openpyxl>=3.1.5",
  "watchdog>=4.0.2",
  "PyPDF2>=3.0.1"
]

[project.optional-dependencies]
dev = ["pytest>=8.2.2", "ruff>=0.6.9"]

[project.scripts]
contractor = "contractor.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/contractor"]
